+++ packages/nimble-components/src/table/cell/template.ts
