+++ packages/nimble-components/src/table/row/index.ts
@@ -2,7 +2,7 @@ import { observable } from '@microsoft/fast-element';
 import { DesignSystem, FoundationElement } from '@microsoft/fast-foundation';
 import { styles } from './styles';
 import { template } from './template';
-import type { TableFieldValue, TableRecord } from '../../types';
+import type { TableFieldValue, TableRecord } from '../types';
 
 declare global {
     interface HTMLElementTagNameMap {
