+++ packages/nimble-components/src/table/row/template.ts
